requirements.txt
accelerate==0.33.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.10
fastapi==0.116.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.1.1
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.11.2
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pillow==10.4.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.5.1
scipy==1.16.1
sentence-transformers==3.0.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
streamlit==1.37.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.66.4
transformers==4.42.4
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.1
watchdog==4.0.2
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
